+++ R/core_solve.R
@@ -2,7 +2,7 @@
 
 #' Calculate the PDFs
 #'
-#' This method takes the a model, the time and space vectors, and the
+#' This method takes a model, the time and space vectors, and the
 #' unpacked parameters for solving the PDF to derive the first passage
 #' time across all conditions. It is a wrapper around the cpp implementations
 #' and [dRiftDM::add_residual]. Important: This function is used in the